add date extension for getting interval date.
